bigquery-aggregation/aggregate_sensor_data/requirements.txt
google-cloud-bigquery==2.29.0
Flask==2.3.1